pyproject.toml
[tool.poetry]
name = "Compsio"
version = "0.1.0"
description = "Lyzr agent with Composio"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]

python = ">=3.11,<3.12"
lyzr-automata = "^0.1.3"
composio-lyzr = "^0.2.22"
composio-core = "^0.2.22"
flask = "^3.0.3"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
serhant = "serhant.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"